searcher/requirements.txt
faiss-cpu==1.13.2
fastapi==0.135.3
uvicorn==0.44.0
httpx==0.28.1
boto3==1.42.85
